minor tests cleanup
